Adds relaunch button in layout preference view
